remove s3 code
