+++ src/main/java/no/nav/kontantstotte/dokument/DokumentService.java
@@ -1,7 +1,5 @@
 package no.nav.kontantstotte.dokument;
 
-import no.nav.kontantstotte.storage.Storage;
-import no.nav.kontantstotte.storage.attachment.AttachmentStorage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,23 +9,17 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 
-import static no.nav.kontantstotte.innlogging.InnloggingUtils.hentFnrFraToken;
-
 @Component
 public class DokumentService {
 
     private static final Logger log = LoggerFactory.getLogger(DokumentService.class);
 
-    private final Storage storage;
-
     private final int maxFileSize;
 
     private FamilieDokumentClient familieDokumentClient;
 
-    DokumentService(AttachmentStorage storage,
-                    @Value("${attachment.max.size.mb}") int maxFileSizeMB,
+    DokumentService(@Value("${attachment.max.size.mb}") int maxFileSizeMB,
                     FamilieDokumentClient familieDokumentClient) {
-        this.storage = storage;
         this.maxFileSize = maxFileSizeMB * 1000 * 1000;
         this.familieDokumentClient = familieDokumentClient;
     }
@@ -55,10 +47,7 @@ public class DokumentService {
             return dokument;
         } else {
             log.info("Feil med hent av vedlegg fra familie-dokument");
+            return null;
         }
-        //TODO: fjerner å hent fra s3 når alle dokument funnes in familie-dokument
-        String directory = hentFnrFraToken();
-        byte[] data = storage.get(directory, key).orElse(null);
-        return data;
     }
 }
