+++ src/main/java/no/nav/kontantstotte/innsending/InnsendingConfiguration.java
@@ -2,7 +2,6 @@ package no.nav.kontantstotte.innsending;
 
 
 import no.nav.kontantstotte.dokument.DokumentService;
-import no.nav.kontantstotte.storage.encryption.EncryptedStorage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,12 +9,10 @@ import org.springframework.context.annotation.Import;
 
 import no.nav.kontantstotte.client.RestClientConfiguration;
 import no.nav.kontantstotte.innsending.oppsummering.OppsummeringConfiguration;
-import no.nav.kontantstotte.storage.s3.S3StorageConfiguration;
 
 @Configuration
 @Import({RestClientConfiguration.class,
-        OppsummeringConfiguration.class,
-        S3StorageConfiguration.class
+        OppsummeringConfiguration.class
 })
 public class InnsendingConfiguration {
