--- src/main/java/no/nav/kontantstotte/storage/StorageException.java
@@ -1,17 +1,0 @@
-package no.nav.kontantstotte.storage;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
-public class StorageException extends RuntimeException {
-
-    public StorageException(String message) {
-        super(message);
-    }
-
-    public StorageException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
