--- src/main/java/no/nav/kontantstotte/storage/attachment/AttachmentConversionException.java
@@ -1,15 +1,0 @@
-package no.nav.kontantstotte.storage.attachment;
-
-import no.nav.kontantstotte.storage.StorageException;
-
-public class AttachmentConversionException extends StorageException {
-
-    public AttachmentConversionException(String message) {
-        super(message);
-    }
-
-    public AttachmentConversionException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-}
